Compile model and fit to data

Compile model and fit to data

fm_fit_model(
  fm_data_inputs,
  SE = 1,
  data_source = 1,
  data_obs = 2,
  samp_process = 0,
  b_constraint = 2,
  const_spphab = 1,
  cov_samp_process = 0,
  biomass_temporal = 1,
  sampling_temporal = 0,
  anisotropy = 0,
  lf_link = 0,
  ref_data = "com",
  EM = "est_b",
  month_ref = 1,
  compute_sd = FALSE
)

Arguments

fm_data_inputs

list Named list obtained from `fm_load_data()`

SE

integer run ADREPORT - 0: no, 1: yes

data_source

integer 1: integrated model (scientific + commercial data), 2: scientific model, 3: commercial model

data_obs

integer observation model for biomass - 1: zero_inflated gamma, 2: zero-inflated lognormal, 3: lognormal

samp_process

integer Sampling process - 0: no sampling process, 1: inhomogeneous Poisson point process

b_constraint

integer put constraint on b parameters - 1: b are positive, 2: no constraints

const_spphab

integer Species-habitat relationship - 1: constant in time

cov_samp_process

integer covariate in the sampling process - 0: none, 1: covariate in the samplign process

biomass_temporal

integer Account for temporal correlation in biomass - 0: no, 1: AR1

sampling_temporal

integer Account for temporal correlation in sampling process - 0: no, 1: AR1

anisotropy

integer Account for anisotropy

lf_link

integer link function of the latent field - 0: log (biomass data), 1: logit (presence absence data)

ref_data

character reference data - "com": commercial (default) or "sci": scientific

EM

character "est_b": b is estimated, "fix_b": b is fixed

month_ref

integer reference month (default is 1 for January)

compute_sd

logical compute SD report (time-consumming)

Value

list A named list of all necessary outputs for generating graphs (step 3)