Load and prepare data for model fitting

Load and prepare data for model fitting

fm_load_data(
  species = "Solea_solea",
  fleet = c("OTB_DEF_>=70_0", "OTB_CEP_>=70_0", "OTT_DEF_>=70_0"),
  survey_data_file = "survey_data.Rdata",
  vmslogbook_data_file = "vmslogbook_data.Rdata",
  study_domain_file = "study_domain.Rdata",
  year_start = 2018,
  year_end = 2018,
  month_start = 11,
  month_end = 11,
  time_step = "Month"
)

Arguments

species

character Name of the species of interest, match the data folder name

fleet

character Vector of conditional operation to filter fleets, the first one will be taken as reference level in the model

survey_data_file

character Name of the Rdata file storing scientific observations

vmslogbook_data_file

character Name of the Rdata file storing vms x logbooks observations

study_domain_file

character Name of the Rdata file storing study domain informations

year_start

integer Lower bound year filter to select data

year_end

integer Upper bound year filter to select data

month_start

integer Lower bound month filter to select data

month_end

integer Upper bound month filter to select data

time_step

character Time intervals, must be Month or Quarter

Value

list A named list of all necessary outputs for model fitting (step 2)

Examples

# run part1
fm_data_inputs <- fm_load_data()
#> Running step 1 -loading data-
#> as(<dgCMatrix>, "dgTMatrix") is deprecated since Matrix 1.5-0; do as(., "TsparseMatrix") instead
#> Joining with `by = join_by(Year, Month)`
#> Joining with `by = join_by(Month, Year, Year_Month)`
#> Joining with `by = join_by(layer, cell)`
#> Joining with `by = join_by(layer, cell)`
#> Joining with `by = join_by(layer, cell)`
#> Step 1 -loading data-: 15.206 sec elapsed