Load and prepare data for model fitting

Load and prepare data for model fitting

fm_load_data(
  species,
  fleet,
  fitted_data = c("biomass", "presabs"),
  survey_data_file,
  vmslogbook_data_file,
  study_domain_file,
  year_start,
  year_end,
  month_start,
  month_end,
  time_step = c("Month", "Quarter"),
  k,
  grid_xmin,
  grid_xmax,
  grid_ymin,
  grid_ymax,
  seed = 29510
)

Arguments

species

character Species of interest

fleet

character Fleet chosen according to the species of interest. A fleet is considered to have homogeneous catchability and targeting behavior.

fitted_data

character Type of the data to be fitted to the model (either `biomass` for biomass data - positive-continuous data - or `presabs` for presence-absence data). Default is `biomass`

survey_data_file

character File containing the scientific (survey) data

vmslogbook_data_file

character File containing the commercial (vmslogbook) data

study_domain_file

character File containing the data describing the study area

year_start

integer Starting year

year_end

integer Ending year

month_start

integer Starting month

month_end

integer Ending month

time_step

character Time step for the model (either `Month` for monthly time step or `Quarter` for quarterly time step). Default is `Month`

k

numeric Parameter controlling the number of knots of the mesh. The higher the denser.

grid_xmin, grid_xmax, grid_ymin, grid_ymax

numeric Limitation of the grid for the spatial domain

seed

integer The seed controlling for random effect. Default is 29510.

Value

list A named list of all necessary outputs for model fitting (step 2)